refactored transaction_context into wallet
